Rename Sirupsen/logrus to sirupsen/logrus
As indicated at the top of the README.md in the
github.com/sirupsen/logrus package, there are issues with
case-sensitivity for this package -- most notably on MacOS where the
HFS+ filesystem is case-insensitive by default. This produces problems
when locally there is a directory called "Sirupsen" and one called
"sirupsen", especially if you are syncing files between machines.

https://github.com/sirupsen/logrus/blob/master/README.md shows that the
recommended spelling is the one that this change makes.